feat: 중복체크 후 중복이면 모달띄우기
